UK Lottery Number Picker
Based on the UK lottery results since it started, this uses the 20 most occurring lottery numbers and randomly selects 6 of the numbers that you can use on the lottery.

When I initially put it together over a year ago it did not have a front end UI it just generated the numbers within the code terminal. I had always liked the Zoltar wish granting carnival machine that was on the Tom Hanks movie 'Big' so I just had to have it as part of the s=design.

A little bit of prompt engineering and a little AI know how and PyGame later and I came up with a version of the Zoltar machine as the main image people see and the balls generate the numbers. To add a little extra magic included carnival sounds and music together with an evil laugh to help set the scene and make using generator more interesting.

The next in the development will be adding animation to the machine to simulate a coin being used and the numbers actually rolling out of the machine one by one.

The numbers themselves will also be selected from the numbers that have occured most regularly together. Or at least it will provide it as an option.